requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.20.0
streamlit
pandas
numpy
plotly
python-dotenv
pyyaml
bcrypt
google-genai
openpyxl
google-generativeai
python-doten